backend_server/src/requirements.txt
astroid==2.12.10
attrs==22.1.0
Automat==20.2.0
azapi==3.0.8
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
constantly==15.1.0
dill==0.3.5.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
incremental==22.10.0
isort==5.10.1
joblib==1.2.0
klein==21.8.0
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
nltk==3.7
platformdirs==2.5.2
pylint==2.15.3
pymusixmatch==0.3
python-dotenv==0.21.0
regex==2022.9.13
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
tomlkit==0.11.5
tqdm==4.64.1
Tubes==0.2.1
Twisted==22.8.0
twisted-iocpsupport==1.0.2
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
zope.interface==5.5.0
